requirements.txt
fastapi
uvicorn
pytube
openai-whisper
youtube_transcript_api
